+++ CHANGELOG.markdown
@@ -3,9 +3,10 @@
  * [Fix #41][i41] — Add an "updated but not restarted" info message (Ruby backend).
  * Bump Bootstrap, CodeMirror, Esprima and Hogan.js vendor libraries — faster parsing, highlighting and rendering!
  * Update to latest versions of jQuery and Underscore.js.
- * API now errors when trying to update an `_id` property.
+ * [Fix #64][i64] — Show an error message when trying to update an `_id` property.
 
 [i41]: https://github.com/bobthecow/genghis/issues/41
+[i64]: https://github.com/bobthecow/genghis/issues/64
 
 
 ## v2.1.6